+++ src/tests/routes/private/mod.rs
@@ -1,0 +1,1 @@
+mod crate_owner_invitations;
